Update Swagger title to reflect Submissions API focus
The `SwaggerExtensions` class was updated to modify the title
of the Swagger documentation. In the `AddSwaggerDocumentation`
method, the `SwaggerDoc` configuration was changed to update
the `Title` property of the `OpenApiInfo` object from
"GradingSystem.Users" to "GradingSystem.Submissions".

This change aligns the Swagger documentation with the new
focus on the Submissions API.